# shared-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Veresiye Platform API
  version: 0.1.0
  description: |
    The single contract app-pos, app-mobile and the backend all speak (see
    veresiye-platform-tasarim.md §4.1). Generated clients / a Prism mock come from
    this file. Source of truth for endpoints/bodies: docs/api-endpoints.md; for the
    data shapes: docs/db-schema.md.

    Conventions:
      - Money: every *_minor is an integer in minor units (kuruş). 50 TL = 5000. No floats.
      - Scaling: orderBody item quantity and tax_percent are ×1000 (1000 = 1 unit / 10%).
      - seller_id / buyer are taken from the bearer token on scoped endpoints, not the body.
      - Field names are snake_case on the wire; Kotlin domain is camelCase (mapper converts).

    Tags:
      - "active"  = needed now (both clients' real mock behaviour). Fully modelled.
      - "future"  = planned but not wired yet (schema/contract placeholder; code skeleton
                    lands in Room but stays unconnected — see plan Aşama 3).
servers:
  - url: http://127.0.0.1:4010
    description: Prism mock (npx @stoplight/prism-cli mock openapi.yaml)
  - url: https://api.veresiye.example
    description: Backend (later phase)
...
security:
  - bearerAuth: []
paths:
...
  /transactions:
    post:
      tags: [transactions]
      summary: Append a ledger entry (single write point; idempotent)
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Equals transaction_id; a retry returns the same entry.
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionCreate' }
      responses:
        '201':
          description: Written
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '200':
          description: Idempotent replay (same entry)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
    get:
      tags: [transactions]
      summary: A customer's history with the seller (newest first)
      parameters:
        - { name: customer_id, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Transaction' }
              # Ahmet's (c1) ledger from the seed: 50 + 30 − 40 = 40,00 TL (newest first).
              example:
                - { transaction_id: t3, seller_id: u_owner, customer_id: c1, amount_minor: 4000, type: PAYMENT, description: Nakit ödeme, basket_id: null, settled_via_pgw: false, receipt_no: null, created_at: '2026-07-22T15:00:00Z' }
                - { transaction_id: t2, seller_id: u_owner, customer_id: c1, amount_minor: 3000, type: DEBT, description: Peynir, basket_id: null, settled_via_pgw: false, receipt_no: null, created_at: '2026-07-21T07:40:00Z' }
                - { transaction_id: t1, seller_id: u_owner, customer_id: c1, amount_minor: 5000, type: DEBT, description: 'Ekmek, süt', basket_id: null, settled_via_pgw: false, receipt_no: null, created_at: '2026-07-20T06:15:00Z' }
  /balances:
    get:
      tags: [transactions]
      summary: The (seller, customer) balance
      parameters:
        - { name: customer_id, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Balance' }

  # ----------------------------------------------------------------- buyer-scoped
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    TransactionType:
      type: string
      enum: [DEBT, PAYMENT]
...
    OrderItem:
      type: object
      required: [name, price, quantity, tax_percent]
      properties:
        name: { type: string }
        price:
          type: integer
          format: int64
          description: 'per-unit, minor units (kuruş)'
        quantity:
          type: integer
          format: int64
          description: 'scaled ×1000 (1000 = 1 unit)'
        tax_percent:
          type: integer
          format: int64
          description: 'scaled ×1000 (1000 = 10%)'
        section_no: { type: integer, default: 1 }
        status: { type: integer, default: 1 }
        type: { type: integer, default: 0 }
        item_limit: { type: integer, format: int64, default: 0 }
    OrderBody:
      type: object
      required: [basket_id, items]
      properties:
        basket_id: { type: string, format: uuid }
        create_invoice: { type: boolean, default: false }
        document_type: { type: integer, default: 0 }
        is_void: { type: boolean, default: false }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }

    # --- transaction ---
    Transaction:
      type: object
      required: [transaction_id, seller_id, customer_id, amount_minor, type, description, settled_via_pgw, created_at]
      # Mirrors FakeRepository seed t1 (Ahmet's first debt, 50,00 TL).
      example:
        transaction_id: t1
        seller_id: u_owner
        customer_id: c1
        amount_minor: 5000
        type: DEBT
        description: Ekmek, süt
        basket_id: null
        basket: null
        settled_via_pgw: false
        receipt_no: null
        created_at: '2026-07-20T06:15:00Z'
      properties:
        transaction_id: { type: string, format: uuid }
        seller_id: { type: string }
        customer_id: { type: string }
        amount_minor: { type: integer, format: int64, minimum: 0 }
        type: { $ref: '#/components/schemas/TransactionType' }
        description: { type: string }
        basket_id: { type: string, format: uuid, nullable: true }
        # The basket itself, not only its id. Nothing serves a basket on its own, so an
        # id without the items beside it is a reference the client cannot resolve.
        basket:
          allOf: [{ $ref: '#/components/schemas/OrderBody' }]
          nullable: true
        settled_via_pgw: { type: boolean, default: false }
        receipt_no: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    TransactionCreate:
      type: object
      required: [transaction_id, customer_id, amount_minor, type, description]
      properties:
        transaction_id: { type: string, format: uuid }
        customer_id: { type: string }
        amount_minor: { type: integer, format: int64, minimum: 0 }
        type: { $ref: '#/components/schemas/TransactionType' }
        description: { type: string }
        basket:
          type: object
          nullable: true
          description: Optional; omit for a money-only entry.
          allOf: [{ $ref: '#/components/schemas/OrderBody' }]
    Balance:
      type: object
      required: [seller_id, customer_id, balance_minor, as_of]
      example:
        seller_id: u_owner
        customer_id: c1
        balance_minor: 4000
        as_of: '2026-07-30T12:00:00Z'
      properties:
        seller_id: { type: string }
        customer_id: { type: string }
        balance_minor: { type: integer, format: int64 }
        as_of: { type: string, format: date-time }
...